FastVaultEmailView refactor
